Columns

+++ segfault-api/src/main/java/heigvd/ch/segfaultapi/repositories/DiscussionRepository.java
@@ -3,5 +3,5 @@ package heigvd.ch.segfaultapi.repositories;
 import heigvd.ch.segfaultapi.model.Discussion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiscussionRepository extends JpaRepository<Discussion, Long> {
+public interface DiscussionRepository extends JpaRepository<Discussion, Integer> {
 }
